+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/ScheduleRestController.java
@@ -39,4 +39,9 @@ public class ScheduleRestController {
     public List<Schedule> getScheduleByClass(@PathVariable Long classId){
         return scheduleController.getScheduleByClass(classId);
     }
+
+    @DeleteMapping("/{scheduleId}") //
+    public void deleteSchedule(@PathVariable Long scheduleId) {
+        scheduleController.deleteSchedule(scheduleId);
+    }
 }