+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/ScheduleController.java
@@ -79,4 +79,8 @@ public class ScheduleController {
         }
         return classSchedule;
     }
+
+    public void deleteSchedule(Long scheduleId) {
+        scheduleService.deleteSchedule(scheduleId);
+    }
 }