+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/ScheduleService.java
@@ -26,4 +26,8 @@ public class ScheduleService {
     public List<Schedule> findAllSchedule() {
         return scheduleRepository.findAll();
     }
+
+    public void deleteSchedule(Long scheduleId) {
+        scheduleRepository.deleteById(scheduleId);
+    }
 }